requirements.txt
fastapi
uvicorn
openai
python-dotenv
pymupdf
fastmcp
gradio